requirements.txt
torch
torchvision
streamlit
opencv-python
numpy
pillow
